+++ bin/main.dart
@@ -25,21 +25,18 @@ void helloworld(Map<String, CoffeeParameter> params) {
   }
 }
 
-main(List<String> args) {
-  CoffeeCli cli = new CoffeeCli("My Cli", [
-    new CoffeeCommand("hello", helloworld, subcommands: [
-      new CoffeeCommand("world", helloworld, parameters: [
-        new CoffeeStringParameter("name", isOptional: false, help: "Use you name", question: "What is your Name ?"),
-        new CoffeeBoolParameter("uppercase",
-            isOptional: false, help: "Big Hello World", question: "Do you want to use uppercase ?"),
-        new CoffeeStringParameter("style",
-            isOptional: true,
-            help: "Style",
-            question: "What kind of style ?",
-            possibleValues: ["CamelCase", "snake_case"])
-      ])
-    ])
-  ]);
+CoffeeCommand get helloCommand => new  CoffeeCommand("hello", helloworld);
+CoffeeCommand get helloWorldCommand =>  new CoffeeCommand("world", helloworld, parameters: [
+  new CoffeeStringParameter("name",  help: "Use you name", question: "What is your Name ?"),
+  new CoffeeBoolParameter("uppercase", help: "Big Hello World", question: "Do you want to use uppercase ?"),
+  new CoffeeStringParameter("style",
+      isOptional: true,
+      help: "Style",
+      question: "What kind of style ?",
+      allowed: ["CamelCase", "snake_case"])
+]);
+CoffeeCommand get helloCommandComplex => new CoffeeCommand("complex", helloworld, commands: [ helloWorldCommand]);
 
-  return cli.execute(args);
+main(List<String> args) {
+  return new CoffeeCli([ helloCommand, helloCommandComplex, helloWorldCommand]).execute(args);
 }