+++ lib/src/cli.dart
@@ -2,53 +2,6 @@ import "dart:io";
 import "package:args/args.dart";
 import "package:ansicolor/ansicolor.dart";
 import "command.dart";
-import 'parameter.dart';
-
-typedef void executeCommand(Map<String, CoffeeParameter> parameters);
-
-class CoffeeCli {
-  final String name;
-  final List<CoffeeCommand> commands;
-  ArgParser _parser;
-
-  CoffeeCli(this.name, this.commands) {
-    _generateArgParser();
-  }
-
-  void _generateArgParser() {
-    _parser = new ArgParser();
-    for (CoffeeCommand cmd in commands) {
-      _parser.addCommand(cmd.name, cmd.parser);
-    }
-  }
-
-  printUsage() {
-    print("$name usage: ");
-    for (CoffeeCommand cmd in commands) {
-      stdout.write("\t${cmd.name} ");
-      print("\t${cmd.parser.usage.split("\n").join("\n\t\t")}");
-    }
-  }
-
-  execute(List<String> args) {
-    try {
-      ArgResults results = _parser.parse(args);
-
-      if (results.command != null) {
-        CoffeeCommand cmd =
-            commands.firstWhere((CoffeeCommand _) => _.name == results.command.name, orElse: () => null);
-        if (cmd == null) {
-          printUsage();
-        }
-        cmd.execute(results.command.arguments);
-      } else {
-       printUsage();
-      }
-    } on FormatException catch (_) {
-      printUsage();
-    }
-  }
-}
 
 String outputRed(String msg, {bool bold}) {
   AnsiPen pen = new AnsiPen()..red(bold: bold);
@@ -63,4 +16,55 @@ String outputWhite(String msg, {bool bold}) {
 String outputGreen(String msg, {bool bold}) {
   AnsiPen pen = new AnsiPen()..green(bold: bold);
   return pen(msg);
+}
+
+String outputBlue(String msg, {bool bold}) {
+  AnsiPen pen = new AnsiPen()..blue(bold: bold);
+  return pen(msg);
+}
+
+
+
+class CoffeeCli {
+
+  final List<CoffeeCommand> commands;
+  final ArgParser parser;
+
+  CoffeeCli(this.commands) : parser = new ArgParser() {
+    for (CoffeeCommand cmd in commands) {
+      parser.addCommand(cmd.name, cmd.parser);
+    }
+    parser.addFlag("help", abbr: "h", help: "Print Usage", negatable: false);
+  }
+
+  printUsage() {
+    print("Usage:");
+    for (CoffeeCommand cmd in commands) {
+      cmd.printUsage();
+    }
+  }
+
+
+  int execute(List<String> args) {
+    try {
+     ArgResults results = parser.parse(args);
+
+      if (results.command != null) {
+        CoffeeCommand cmd = commands.firstWhere((CoffeeCommand _) => _.name == results.command.name, orElse: () => null);
+        if (cmd == null) {
+          cmd.printUsage();
+        }
+        return cmd.execute(results.command.arguments);
+      } else {
+        printUsage();
+      }
+    } on FormatException catch (e) {
+      stderr.writeln(outputRed(e.message));
+      printUsage();
+    } catch (_, stacktrace) {
+      print(_);
+      //print(stacktrace);
+    }
+    return 1;
+  }
 }