+++ CHANGELOG.md
@@ -2,7 +2,14 @@
 
 ## 0.2.0
 
-Can add CoffeeCommand as subcommand of an other CoffeeCommand
+* Can add CoffeeCommand as subcommand of an other CoffeeCommand
+* Better usage
+
+###Breaking Change
+- `possibleValues` in `CoffeeParameter` change by `allowed`
+- `CoffeeCommand`
+    - `command` rename by `executable`
+    - `subcommands` rename by `commands`
 
 ## 0.1.1